Simplify & shorten additional posixlib socket code paths
