+++ posixlib/src/main/resources/scala-native/netdb.h
@@ -4,11 +4,6 @@
 #else
 #include <netdb.h>
 #endif
-#include "sys/socket_conversions.h"
-
-#ifndef __SYS_SOCKET_H
-#include "sys/socket.h"
-#endif
 
 struct scalanative_addrinfo {
     int ai_flags;                         /* Input flags.  */
