+++ posixlib/src/main/scala/scala/scalanative/posix/sys/socket.scala
@@ -12,6 +12,23 @@ object socket {
   type _14 = Nat.Digit2[Nat._1, Nat._4]
   type _15 = Nat.Digit2[Nat._1, Nat._5]
 
+  /* Design Note:
+   *  C11 _Static_assert() statements in socket.c check that the
+   *  Scala Native structures declared below match, closely enough for
+   *  purpose, the corresponding structure declared by the operating
+   *  system.
+   *
+   *  The transcription from Scala declarations here to the corresponding
+   *  'scalanative_foo' declarations is manual, not automatic.
+   *  If you change a declaration here, please also check/add/delete
+   *  the C declaration & checks.
+   *
+   *  Keeping Scala & operating system structures synchronized allows
+   *  the vast majority of structures to be passed to & from without
+   *  needing an expensive "glue" conversion layer.  The C compiler
+   *  does the work once at compilation, rather than at each runtime call.
+   */
+
   type socklen_t = CUnsignedInt
 
   type sa_family_t = CUnsignedShort
@@ -21,11 +38,9 @@ object socket {
     CArray[CChar, _14] // sa_data, size = 14 in OS X and Linux
   ]
 
-  /* The declaration of sockaddr_storage should yield 128 bytes.
-   * Code in socket.c checks that OS sockaddr_storage is that size.
-   * If/when changing, keep the two in correspondence.
+  /* The declaration of sockaddr_storage should yield 128 bytes,
+   * with an overall alignment so that pointers have natural (64 )alignment.
    */
-
   type sockaddr_storage = CStruct4[
     sa_family_t, // ss_family, // ss_family, sa_len is synthisized if needed
     CUnsignedShort, // __opaquePadTo32
@@ -172,8 +187,23 @@ object socket {
    * layer to handle required conversions. Usually the structure
    * in question is a sockaddr or pointer to one.
    */
+  /* Design Note:
+   *   Most of these are fast, direct call to C. See 'Design Note' at
+   *   top of this file.
+   *
+   *   Scalar/primitive values passed directly to/from the
+   *   operating system (OS) cause no problems.
+   *
+   *   When one has assurance from the C compiler that the Scala Native
+   *   that Scala Native structures match, closely enough, the corresponding
+   *   structure used by the operating system, one can use the former
+   *   as arguments in direct calls to the OS.
+   *
+   *   Scala methods which need their arguments transformed use a
+   *   '@name' annotation. See (Pending future) methods 'sendmsg()' &
+   *   'recvmsg'.
+   */
 
-  @name("scalanative_getsockname")
   def getsockname(
       socket: CInt,
       address: Ptr[sockaddr],
@@ -182,20 +212,17 @@ object socket {
 
   def socket(domain: CInt, tpe: CInt, protocol: CInt): CInt = extern
 
-  @name("scalanative_connect")
   def connect(
       socket: CInt,
       address: Ptr[sockaddr],
       address_len: socklen_t
   ): CInt = extern
 
-  @name("scalanative_bind")
   def bind(socket: CInt, address: Ptr[sockaddr], address_len: socklen_t): CInt =
     extern
 
   def listen(socket: CInt, backlog: CInt): CInt = extern
 
-  @name("scalanative_accept")
   def accept(
       socket: CInt,
       address: Ptr[sockaddr],
@@ -225,7 +252,6 @@ object socket {
       flags: CInt
   ): CSSize = extern
 
-  // direct call to C, _Static_assert in socket.c validates structures.
   def recvfrom(
       socket: CInt,
       buffer: Ptr[Byte],
@@ -242,7 +268,6 @@ object socket {
       flags: CInt
   ): CSSize = extern
 
-  // direct call to C, _Static_assert in socket.c validates structures.
   def sendto(
       socket: CInt,
       buffer: Ptr[Byte],
@@ -326,6 +351,7 @@ object socketOps {
     def msg_control: Ptr[Byte] = ptr._5
     def msg_controllen: socklen_t = ptr._6
     def msg_flags: CInt = ptr._7
+
     def msg_name_=(v: Ptr[Byte]): Unit = ptr._1 = v
     def msg_namelen_=(v: socklen_t): Unit = ptr._2 = v
     def msg_iov_=(v: Ptr[uio.iovec]): Unit = ptr._3 = v
@@ -339,6 +365,7 @@ object socketOps {
     def cmsg_len: socklen_t = ptr._1
     def cmsg_level: CInt = ptr._2
     def cmsg_type: CInt = ptr._3
+
     def cmsg_len_=(v: socklen_t): Unit = ptr._1 = v
     def cmsg_level_=(v: CInt): Unit = ptr._2 = v
     def cmsg_type_=(v: CInt): Unit = ptr._3 = v
@@ -350,6 +377,7 @@ object socketOps {
 
     def l_onoff: CInt = if (isWindows) asWinLinger._1.toInt else ptr._1
     def l_linger: CInt = if (isWindows) asWinLinger._2.toInt else ptr._2
+
     def l_onoff_=(v: CInt): Unit =
       if (isWindows) asWinLinger._1 = v.toUShort
       else ptr._1 = v
